Refactor de Extra y save evento

+++ src/main/kotlin/com/estonianport/agendaza/dto/CantidadesPanelAdminDTO.kt
@@ -3,15 +3,15 @@ package com.estonianport.agendaza.dto
 import java.io.Serializable
 
 class CantidadesPanelAdminDTO(
-    var cantUsuarios: Long = 0,
-    var cantTipoEvento: Long = 0,
-    var cantExtras: Long = 0,
-    var cantPagos: Long = 0,
-    var cantEventos: Long = 0,
-    var cantCliente: Long = 0,
-    var cantCatering: Long = 0,
-    var cantServicios: Long = 0,
-    var cantClausula: Long = 0
+    var cantUsuarios: Int = 0,
+    var cantTipoEvento: Int = 0,
+    var cantExtras: Int = 0,
+    var cantPagos: Int = 0,
+    var cantEventos: Int = 0,
+    var cantCliente: Int = 0,
+    var cantCatering: Int = 0,
+    var cantServicios: Int = 0,
+    var cantClausula: Int = 0
 ) : Serializable {
 
     companion object {
