+++ src/main/kotlin/com/estonianport/agendaza/repository/ServicioRepository.kt
@@ -43,5 +43,5 @@ interface ServicioRepository : CrudRepository<Servicio, Long> {
     fun getAllServicioByTipoEventoId(tipoEventoId: Long): List<ServicioDTO>
 
     @Query("SELECT COUNT(s) FROM Servicio s JOIN s.listaEmpresa e WHERE e.id = :id AND s.fechaBaja IS NULL")
-    fun countActivosByEmpresaId(id: Long): Long
+    fun countActivosByEmpresaId(id: Long): Int
 }
